Extract MapperRequest out to mapper package
* Reuse MapperRequest in OrderSummaryConverter.
