+++ src/test/order_item_summary/OrderItemSummaryFactory.unit.test.ts
@@ -6,7 +6,7 @@ import { mockCertificateItem, mockMissingImageDeliveryItem } from "../__mocks__/
 import { expect } from "chai";
 import { MissingImageDeliveryMapper } from "../../order_item_summary/MissingImageDeliveryMapper";
 import { NullOrderItemMapper } from "../../order_item_summary/NullOrderItemMapper";
-import { MapperRequest } from "../../order_item_summary/MapperRequest";
+import { MapperRequest } from "../../mappers/MapperRequest";
 
 const sandbox = sinon.createSandbox();
